Add system role ordering

+++ src/Database/Tables/SystemRoleTable.php
@@ -3,9 +3,11 @@ declare(strict_types = 1);
 
 namespace Database\Tables;
 
+use Data\UserId;
 use Database\AbstractTable;
 use Database\Objects\RoleInfo;
 use Exception;
+use LogicException;
 use PDOException;
 
 final class SystemRoleTable extends AbstractTable{
@@ -19,18 +21,27 @@ final class SystemRoleTable extends AbstractTable{
     $this->db->beginTransaction();
     
     try{
-      $this->execute('INSERT INTO system_roles (title, special) VALUES (?, ?)',
-                     'SB', [$title, $special]);
+      $sql = <<<SQL
+SELECT IFNULL(MAX(id) + 1, 1)       AS id,
+       IFNULL(MAX(ordering) + 1, 1) AS ordering
+FROM system_roles
+SQL;
       
-      $id = $this->getLastInsertId();
+      $stmt = $this->db->query($sql);
+      $next = $this->fetchOneRaw($stmt);
       
-      if ($id === null){
+      if ($next === false){
         $this->db->rollBack();
-        throw new Exception('Could not retrieve role ID.');
+        throw new LogicException('Error calculating next role ID.');
       }
       
+      $role_id = $next['id'];
+      
+      $this->execute('INSERT INTO system_roles (id, title, ordering, special) VALUES (?, ?, ?, ?)',
+                     'ISIB', [$role_id, $title, $special ? 0 : $next['ordering'], $special]);
+      
       $this->db->commit();
-      return $id;
+      return $role_id;
     }catch(PDOException $e){
       $this->db->rollBack();
       throw $e;
@@ -42,12 +53,65 @@ final class SystemRoleTable extends AbstractTable{
                    'SI', [$title, $id]);
   }
   
+  public function swapRolesIfNotSpecial(int $ordering1, int $ordering2): void{
+    $sql = <<<SQL
+UPDATE system_roles sr1 INNER JOIN system_roles sr2 ON sr1.ordering = ? AND sr2.ordering = ?
+SET sr1.ordering = sr2.ordering,
+    sr2.ordering = sr1.ordering
+WHERE sr1.special = FALSE AND sr2.special = FALSE
+SQL;
+    
+    $this->execute($sql, 'II', [$ordering1, $ordering2]);
+  }
+  
+  public function findMaxOrdering(): ?int{
+    $stmt = $this->db->query('SELECT MAX(ordering) FROM system_roles');
+    return $this->fetchOneInt($stmt);
+  }
+  
+  public function isRoleAssignableBy(int $role_id, UserId $user_id): bool{
+    $sql = <<<SQL
+SELECT 1
+FROM system_roles sr
+WHERE id = ?
+  AND special = FALSE
+  AND ordering > IFNULL((SELECT ordering
+                         FROM system_roles sr2
+                         JOIN users u ON sr2.id = u.role_id
+                         WHERE u.id = ?), ~0)
+SQL;
+    
+    $stmt = $this->execute($sql, 'IS', [$role_id, $user_id]);
+    return $this->fetchOneInt($stmt) !== null;
+  }
+  
   /**
    * @return RoleInfo[]
    */
   public function listRoles(): array{
-    $stmt = $this->db->query('SELECT id, title, special FROM system_roles ORDER BY special DESC, id ASC');
-    return $this->fetchMap($stmt, fn($v): RoleInfo => new RoleInfo($v['id'], $v['title'], 0, (bool)$v['special']));
+    $stmt = $this->db->query('SELECT id, title, ordering, special FROM system_roles ORDER BY special DESC, ordering ASC');
+    return $this->fetchMap($stmt, fn($v): RoleInfo => new RoleInfo($v['id'], $v['title'], $v['ordering'], (bool)$v['special']));
+  }
+  
+  /**
+   * @param UserId $user_id
+   * @return RoleInfo[]
+   */
+  public function listRolesAssignableBy(UserId $user_id): array{
+    $sql = <<<SQL
+SELECT id, title, ordering, special
+FROM system_roles sr
+WHERE special = FALSE
+  AND ((SELECT u.admin FROM users u WHERE u.id = ?) OR
+       (ordering > IFNULL((SELECT ordering
+                           FROM system_roles sr2
+                           JOIN users u ON sr2.id = u.role_id
+                           WHERE u.id = ?), ~0)))
+ORDER BY ordering ASC
+SQL;
+    
+    $stmt = $this->execute($sql, 'SS', [$user_id, $user_id]);
+    return $this->fetchMap($stmt, fn($v): RoleInfo => new RoleInfo($v['id'], $v['title'], (int)$v['ordering'], (bool)$v['special']));
   }
   
   public function getRoleIdByTitle(string $title): ?int{
@@ -65,8 +129,30 @@ final class SystemRoleTable extends AbstractTable{
   }
   
   public function deleteById(int $id): void{
-    $this->execute('DELETE FROM system_roles WHERE id = ? AND special = FALSE',
-                   'I', [$id]);
+    $this->db->beginTransaction();
+    
+    try{
+      $stmt = $this->execute('SELECT ordering FROM system_roles WHERE id = ? AND special = FALSE',
+                             'I', [$id]);
+      
+      $ordering = $this->fetchOneInt($stmt);
+  
+      if ($ordering === null){
+        $this->db->rollBack();
+        return;
+      }
+      
+      $this->execute('UPDATE system_roles SET ordering = ordering - 1 WHERE ordering > ? AND special = FALSE',
+                     'I', [$ordering]);
+      
+      $this->execute('DELETE FROM system_roles WHERE id = ? AND special = FALSE',
+                     'I', [$id]);
+      
+      $this->db->commit();
+    }catch(PDOException $e){
+      $this->db->rollBack();
+      throw $e;
+    }
   }
 }
